# Faster string serialization