I've applied the @addaleax's suggestions (#23) and we got a sensible increase of the performances in string serialization! :)

Before
```
JSON.stringify short string x 5,045,063 ops/sec ±1.14% (88 runs sampled)
fast-json-stringify short string x 11,559,636 ops/sec ±0.82% (90 runs sampled)
```
After
```
JSON.stringify short string x 5,088,404 ops/sec ±0.79% (92 runs sampled)
fast-json-stringify short string x 12,171,333 ops/sec ±0.75% (90 runs sampled)
```

Nice catch @addaleax! :)
cc @mcollina